policy summary assistnt/requirements.txt
streamlit==1.38.0

langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-openai==0.3.5
langchain-text-splitters==0.3.2

chromadb==0.4.24
sentence-transformers==2.6.1

torch==2.1.2
transformers==4.41.1

pdfplumber==0.10.4
python-dotenv==1.0.1
